+++ src/data.py
@@ -1,8 +1,12 @@
 RULES = {
         "add": [
         {
-            "value": """meme has:images -is:retweet -is:reply lang:en -crypto -nft -politics""",
+            "value": """(meme OR memes OR humor OR funny) has:images -is:retweet -is:reply lang:en -crypto -nft -politics""",
             "tag": "Funny things"
+        },
+        {
+            "value": """(meme OR memes OR humor OR funny) (tech OR coding OR programming) has:images -is:retweet -is:reply lang:en -crypto -nft -politics""",
+            "tag": "Tech"
         },
     ]
 }